4.5 refactoring utils js - разделяет и властвует
