+++ src/mock/film.js
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import { getRandomNumber, getRandomItem, shuffleArr } from '../utils/utils.js';
+import { getRandomNumber, getRandomItem, shuffleArr } from '../utils/random.js';
 import { getRandomDate } from '../utils/days.js';
 import { COUNTRIES, GENRES } from '../const.js';
 import { getComments } from './comments.js';
