+++ src/view/rank.js
@@ -1,5 +1,5 @@
 import Abstract from './abstract.js';
-import { getRandomNumber } from '../utils/utils.js';
+import { getRandomNumber } from '../utils/random.js';
 
 const viewed = getRandomNumber(0, 31);
